+++ test/functional/test_download_and_import.py
@@ -12,11 +12,13 @@
 # language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
+import importlib
 import shlex
 import subprocess
 import textwrap
 
 import pytest
+import six
 
 from sagemaker_containers.beta.framework import errors, modules
 import test
@@ -24,9 +26,18 @@ import test
 data = ['from distutils.core import setup\n',
         'setup(name="my_test_script", py_modules=["my_test_script"])']
 
-SETUP = test.File('setup.py', data)
+SETUP_FILE = test.File('setup.py', data)
 
-USER_SCRIPT = test.File('my_test_script.py', 'def validate(): return True')
+USER_SCRIPT_FILE = test.File('my_test_script.py', 'def validate(): return True')
+
+REQUIREMENTS_TXT_ASSERT_STR = """
+ ____                   __  __       _.............
+/ ___|  __ _  __ _  ___|  \/  | __ _| | _____ _ __.
+\___ \ / _` |/ _` |/ _ \ |\/| |/ _` | |/ / _ \ '__|
+ ___) | (_| | (_| |  __/ |  | | (_| |   <  __/ |...
+|____/ \__,_|\__, |\___|_|  |_|\__,_|_|\_\___|_|...
+             |___/.................................
+""".replace('.', ' ').strip()  # noqa W605
 
 
 @pytest.fixture(name='user_module_name')
@@ -39,28 +50,63 @@ def erase_user_module():
         pass
 
 
-def test_import_module(user_module_name):
-    user_module = test.UserModule(USER_SCRIPT).add_file(SETUP).upload()
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_FILE).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_FILE)])
+def test_import_module(user_module, user_module_name):
+    user_module.upload()
+
+    module = modules.import_module(user_module.url, user_module_name)
+
+    assert module.validate()
+
+
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_FILE).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_FILE)])
+def test_import_module_with_s3_script(user_module, user_module_name):
+    user_module.upload()
 
     module = modules.import_module(user_module.url, user_module_name, cache=False)
 
     assert module.validate()
 
 
-def test_import_module_with_s3_script(user_module_name):
-    user_module = test.UserModule(USER_SCRIPT).upload()
-
-    module = modules.import_module(user_module.url, user_module_name, cache=False)
-
-    assert module.validate()
-
-
-def test_import_module_with_local_script(user_module_name, tmpdir):
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_FILE).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_FILE)])
+def test_import_module_with_local_script(user_module, user_module_name, tmpdir):
     tmp_code_dir = str(tmpdir)
 
-    test.UserModule(USER_SCRIPT).create_tmp_dir_with_files(tmp_code_dir)
+    user_module.create_tmp_dir_with_files(tmp_code_dir)
 
     module = modules.import_module(tmp_code_dir, user_module_name, cache=False)
+
+    assert module.validate()
+
+
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_FILE).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_FILE)])
+def test_import_module_via_download_and_install(user_module, user_module_name):
+    user_module.upload()
+
+    modules.download_and_install(user_module.url, name=user_module_name, cache=False)
+    module = importlib.import_module(user_module_name)
+    six.moves.reload_module(module)
+
+    assert module.validate()
+
+
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_FILE).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_FILE)])
+def test_import_module_with_s3_script_via_download_and_install(user_module, user_module_name):
+    user_module.upload()
+
+    modules.download_and_install(user_module.url, name=user_module_name, cache=False)
+    module = importlib.import_module(user_module_name)
+    six.moves.reload_module(module)
 
     assert module.validate()
 
@@ -78,49 +124,36 @@ USER_SCRIPT_WITH_REQUIREMENTS = test.File('my_test_script.py', data)
 REQUIREMENTS_FILE = test.File('requirements.txt', 'pyfiglet')
 
 
-def test_import_module_with_s3_script_with_requirements(user_module_name):
-    user_module = test.UserModule(USER_SCRIPT_WITH_REQUIREMENTS).add_file(REQUIREMENTS_FILE).upload()
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_WITH_REQUIREMENTS).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_WITH_REQUIREMENTS)])
+def test_import_module_with_s3_script_with_requirements(user_module, user_module_name):
+    user_module = user_module.add_file(REQUIREMENTS_FILE).upload()
 
     module = modules.import_module(user_module.url, user_module_name, cache=False)
 
-    assert module.say() == """
- ____                   __  __       _.............
-/ ___|  __ _  __ _  ___|  \/  | __ _| | _____ _ __.
-\___ \ / _` |/ _` |/ _ \ |\/| |/ _` | |/ / _ \ '__|
- ___) | (_| | (_| |  __/ |  | | (_| |   <  __/ |...
-|____/ \__,_|\__, |\___|_|  |_|\__,_|_|\_\___|_|...
-             |___/.................................
-""".replace('.', ' ').strip()  # noqa W605
+    assert module.say() == REQUIREMENTS_TXT_ASSERT_STR
 
 
-data = textwrap.dedent("""
-            import file_2
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_WITH_REQUIREMENTS).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_WITH_REQUIREMENTS)])
+def test_import_module_with_requirements_via_download_and_install(user_module, user_module_name):
+    user_module = user_module.add_file(REQUIREMENTS_FILE).upload()
 
+    modules.download_and_install(user_module.url, name=user_module_name, cache=False)
+    module = importlib.import_module(user_module_name)
+    six.moves.reload_module(module)
 
-            def validate():
-                return file_2.IMPORTED
-""")
-
-USER_SCRIPT_WITH_ADDITIONAL_FILE = test.File('my_test_script.py', data)
-
-ADDITIONAL_FILE = test.File('file_2.py', 'IMPORTED = True')
-
-
-def test_import_module_with_s3_script_with_additional_files(user_module_name):
-    user_module = test.UserModule(USER_SCRIPT_WITH_ADDITIONAL_FILE).add_file(ADDITIONAL_FILE).upload()
-
-    module = modules.import_module(user_module.url, user_module_name, cache=False)
-
-    assert module.validate()
+    assert module.say() == REQUIREMENTS_TXT_ASSERT_STR
 
 
 data = ['raise ValueError("this script does not work")']
-
 USER_SCRIPT_WITH_ERROR = test.File('my_test_script.py', data)
 
 
 def test_import_module_with_s3_script_with_error(user_module_name):
-    user_module = test.UserModule(USER_SCRIPT_WITH_ERROR).upload()
+    user_module = test.UserModule(USER_SCRIPT_WITH_ERROR).add_file(SETUP_FILE).upload()
 
     with pytest.raises(errors.ImportModuleError):
         modules.import_module(user_module.url, user_module_name, cache=False)